+++ api/src/validators/params.validator.ts
@@ -26,7 +26,7 @@ const validateParams = (schema: Schema): RequestHandler => {
       await checkSchema(schema, ["params"]).run(req);
       const errors = validationResult(req);
       if (!errors.isEmpty()) {
-        next({ status: "fail", code: 404, data: errors.mapped() });
+        next({ status: "fail", code: 404, errors: errors.mapped() });
       }
 
       next();