+++ api/src/validators/query.validator.ts
@@ -30,7 +30,7 @@ const validateQuery = (schema: Schema): RequestHandler => {
       await checkSchema(schema, ["query"]).run(req);
       const errors = validationResult(req);
       if (!errors.isEmpty()) {
-        next({ status: "fail", code: 404, data: errors.mapped() });
+        next({ status: "fail", code: 404, errors: errors.mapped() });
       }
 
       next();