+++ api/src/validators/validators.ts
@@ -1,5 +1,6 @@
+import validateParams from "./params.validator";
+import validatePost from "./post.validator";
 import validateSignUp from "./signup.validator";
 import validateQuery from "./query.validator";
-import validateParams from "./params.validator";
 
-export { validateParams, validateSignUp, validateQuery };
+export { validateParams, validatePost, validateSignUp, validateQuery };